+++ src/lang/lang_errors.ml
@@ -186,7 +186,8 @@ let report lexbuf f =
       Printf.sprintf "%sine %d, char %d%s"
         (if start.Lexing.pos_fname = "" then "L"
         else
-          Printf.sprintf "File %s, l" (Utils.escape_utf8 start.Lexing.pos_fname))
+          Printf.sprintf "File %s, l"
+            (Utils.quote_utf8_string start.Lexing.pos_fname))
         start.Lexing.pos_lnum
         (start.Lexing.pos_cnum - start.Lexing.pos_bol)
         (if buf = "" then "" else Printf.sprintf " before %S" buf)