+++ src/lang/lang_lexer.ml
@@ -305,18 +305,19 @@ let rec token lexbuf =
                "Parse error: " ^ Sedlexing.Utf8.lexeme lexbuf ))
 
 and read_string c pos buf lexbuf =
+  (* See: https://en.wikipedia.org/wiki/Escape_sequences_in_C *)
   match%sedlex lexbuf with
-    | '\\', '/' ->
-        Buffer.add_char buf '/';
-        read_string c pos buf lexbuf
-    | '\\', '\\' ->
-        Buffer.add_char buf '\\';
+    | '\\', 'a' ->
+        Buffer.add_char buf '\x07';
         read_string c pos buf lexbuf
     | '\\', 'b' ->
         Buffer.add_char buf '\b';
         read_string c pos buf lexbuf
+    | '\\', 'e' ->
+        Buffer.add_char buf '\x1b';
+        read_string c pos buf lexbuf
     | '\\', 'f' ->
-        Buffer.add_char buf '\012';
+        Buffer.add_char buf '\x0c';
         read_string c pos buf lexbuf
     | '\\', 'n' ->
         Buffer.add_char buf '\n';
@@ -327,7 +328,19 @@ and read_string c pos buf lexbuf =
     | '\\', 't' ->
         Buffer.add_char buf '\t';
         read_string c pos buf lexbuf
-    | '\\', '\n', Star skipped -> read_string c pos buf lexbuf
+    | '\\', 'v' ->
+        Buffer.add_char buf '\x0b';
+        read_string c pos buf lexbuf
+    | '\\', '\\' ->
+        Buffer.add_char buf '\\';
+        read_string c pos buf lexbuf
+    | '\\', ('"' | '\'') ->
+        let matched = Sedlexing.Utf8.lexeme lexbuf in
+        Buffer.add_char buf matched.[1];
+        read_string c pos buf lexbuf
+    | '\\', '?' ->
+        Buffer.add_char buf '\x3f';
+        read_string c pos buf lexbuf
     | '\\', 'x', ascii_hex_digit, ascii_hex_digit ->
         let matched = Sedlexing.Utf8.lexeme lexbuf in
         let idx = String.index matched 'x' in
@@ -335,23 +348,25 @@ and read_string c pos buf lexbuf =
         let code = int_of_string (Printf.sprintf "0x%s" code) in
         Buffer.add_char buf (Char.chr code);
         read_string c pos buf lexbuf
-    | '\\', 'o', oct_digit, oct_digit, oct_digit ->
+    | '\\', oct_digit, oct_digit, oct_digit ->
         let matched = Sedlexing.Utf8.lexeme lexbuf in
-        let idx = String.index matched 'o' in
+        let idx = String.index matched '\\' in
         let code = String.sub matched (idx + 1) 3 in
-        let code = int_of_string (Printf.sprintf "0o%s" code) in
+        let code = min 255 (int_of_string (Printf.sprintf "0o%s" code)) in
         Buffer.add_char buf (Char.chr code);
         read_string c pos buf lexbuf
-    | '\\', decimal_digit, decimal_digit, decimal_digit ->
+    | ( '\\',
+        'u',
+        ascii_hex_digit,
+        ascii_hex_digit,
+        ascii_hex_digit,
+        ascii_hex_digit ) ->
         let matched = Sedlexing.Utf8.lexeme lexbuf in
-        let code = String.sub matched 1 3 in
-        let code = int_of_string code in
-        Buffer.add_char buf (Char.chr code);
+        Buffer.add_string buf (Utils.unescape_utf8_char matched);
         read_string c pos buf lexbuf
-    | '\\', ('"' | '\'') ->
-        let matched = Sedlexing.Utf8.lexeme lexbuf in
-        Buffer.add_char buf matched.[1];
-        read_string c pos buf lexbuf
+    (* Multiline string support: some text \
+       Some more text *)
+    | '\\', '\n', Star skipped -> read_string c pos buf lexbuf
     | '\\', any ->
         Printf.printf "Warning: illegal backslash escape in string.\n";
         Buffer.add_string buf (Sedlexing.Utf8.lexeme lexbuf);